apps/ingestion/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-cloud-documentai==2.32.0
google-cloud-pubsub==2.27.1
google-generativeai==0.8.4
python-pptx==1.0.2
pydantic==2.10.3
structlog==24.4.0
httpx==0.28.1
tenacity==9.0.0
Pillow==11.1.0
